Use the same code for vector intrinsics and java default implementation for floating-point math APIs
Background:

This is a followed-up discussion to the Vector API's potential issue of
FP math operations like "SIN/COS": the results maybe different when the
same API is executed with C2 and interpreter on x86 systems, almost
reproductive [1]. The main reason is the C2 compiler intrinsifies these
APIs by calling the Intel's SVML library for x86 platforms on linux and
windows OSs, while the APIs' default implementation calls the relative
scalar java.lang.Math APIs (e.g. Math.sin/cos). The implementations are
all guaranteed to be in 1.0 ulp error bound which follows the API's
specification. And it sounds reasonable that the different implementations
generate different results for the same input. But the result difference
between interpreter and C2 compiler is confusing, which looks like an
issue that should be fixed.

Solution:

To fix this issue, we can change these APIs' default implementation. The
main idea is to use the same implementation for different API
implementation paths (i.e. the API's Java default implementation and
vector intrinsification by C2 compiler).

Here
    are the main changes:
     1) Add a native method "hasNativeImpl()". It is used to check whether
    there is the native vector implementation for the given math op with
    the given vector type on current architecture.
     2) Add a native method "nativeImpl()". It is used to calculate the
    result of the math op by calling the same native vector library as C2
    compiler.
     3) Change the floating point math API's default implementation. It
    checks whether there is the native vector implementation by calling
    "hasNativeImpl()" firstly, and calls "nativeImpl()" if it is true. The
    scalar Math API is called like before if the native library is not
    supported.

    With this change, the APIs' default java implementation uses the same
    algorithm with the hotspot C2 compiler no matther whether the vector math
    library is supported or not. This can make sure the results for the same
    API are the same when it is compiled by interpreter and C2 compiler.

[1] https://mail.openjdk.org/pipermail/panama-dev/2022-August/017372.html

Jira: ENTLLT-5890
Change-Id: Icb572fb612a7458d881f0a7d3393e8f9a2be8d7f
CustomizedGitHooks: yes